apps/yuqing-sentiment/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
pandas==2.1.3
numpy==1.26.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiohttp
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
loguru==0.7.3
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3
httpx==0.25.2
tenacity==8.2.3
chromadb==0.4.18
sentence-transformers==5.1.0
jieba==0.42.1
akshare==1.17.44
